requirements.txt
monai>=1.0.0
pytorch-lightning>=1.7.0
ruamel.yaml>=0.17.21
kornia>=0.6.9
pandas>=1.5.2
pydicom>=2.3.1
segmentation_models_pytorch>=0.3.1
scikit-image>=0.19.3
GDCM>=1.1
pylibjpeg>=1.4.0
pylibjpeg-libjpeg>=1.3.2
scikit-learn>=1.2.0
tensorboard>=2.11.0
matplotlib>=3.6.2
topologylayer@git+https://github.com/bruel-gabrielsson/TopologyLayer.git@309f49dfde8ccf8e771b3546917552e704048b6b
